+++ projects/ngx-bar-rating/src/lib/bar-rating.module.ts
@@ -1,16 +1,23 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
 import { BarRating } from './bar-rating';
-import { BarRatingPipe } from './bar-rating.pipe';
+import { ActiveRating, InactiveRating, FractionRating } from './custom-rating';
+import { BarRatingEffect } from './bar-rating-effect';
 
 @NgModule({
-  declarations: [BarRating, BarRatingPipe],
   imports: [
-    CommonModule,
-    FormsModule
+    BarRating,
+    BarRatingEffect,
+    ActiveRating,
+    InactiveRating,
+    FractionRating
   ],
-  exports: [BarRating]
+  exports: [
+    BarRating,
+    BarRatingEffect,
+    ActiveRating,
+    InactiveRating,
+    FractionRating
+  ]
 })
 export class BarRatingModule {
 }