+++ projects/ngx-bar-rating/src/lib/bar-rating.ts
@@ -1,175 +1,162 @@
 import {
   Component,
-  Input,
-  Output,
-  EventEmitter,
-  OnInit,
-  OnChanges,
-  SimpleChanges,
-  ChangeDetectorRef,
+  inject,
+  signal,
+  output,
+  computed,
+  forwardRef,
+  numberAttribute,
+  booleanAttribute,
+  input,
+  model,
+  contentChild,
+  Signal,
+  Provider,
+  InputSignal,
+  ModelSignal,
+  WritableSignal,
+  OutputEmitterRef,
   ChangeDetectionStrategy,
-  forwardRef
+  InputSignalWithTransform
 } from '@angular/core';
-import { ControlValueAccessor, Validator, NG_VALIDATORS, NG_VALUE_ACCESSOR, FormControl } from '@angular/forms';
+import { NgTemplateOutlet } from '@angular/common';
+import { ControlValueAccessor, Validator, NG_VALIDATORS, NG_VALUE_ACCESSOR, UntypedFormControl } from '@angular/forms';
+import { ActiveRating, FractionRating, InactiveRating } from './custom-rating';
+import { BAR_RATING_OPTIONS, BarRatingOptions } from './bar-rating.model';
 
 /** This allows support [(ngModel)] and ngControl. */
-const RATING_VALUE_ACCESSOR = {
+const RATING_VALUE_ACCESSOR: Provider = {
   provide: NG_VALUE_ACCESSOR,
   useExisting: forwardRef(() => BarRating),
   multi: true
 };
 
 /** This allows control required validation. */
-const RATING_VALUE_VALIDATOR = {
+const RATING_VALUE_VALIDATOR: Provider = {
   provide: NG_VALIDATORS,
   useExisting: forwardRef(() => BarRating),
   multi: true,
 };
 
+enum BarRatingUnitStateEnum {
+  active = 'active',
+  inactive = 'inactive',
+  selected = 'selected',
+  fraction = 'fraction'
+}
+
+type BarRatingUnitState = `${ BarRatingUnitStateEnum }`;
+
 @Component({
   selector: 'bar-rating',
   templateUrl: './bar-rating.html',
-  styleUrls: ['./bar-rating.scss'],
+  styleUrl: './bar-rating.scss',
   providers: [RATING_VALUE_ACCESSOR, RATING_VALUE_VALIDATOR],
+  imports: [NgTemplateOutlet],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class BarRating implements OnInit, OnChanges, ControlValueAccessor, Validator {
+export class BarRating implements ControlValueAccessor, Validator {
 
-  contexts: { fraction: boolean, selected: boolean, active: boolean, click: (e) => void, enter: () => void }[] = [];
-  nextRate: number;
+  readonly defaultOptions: BarRatingOptions = inject(BAR_RATING_OPTIONS);
+
+  onChange: OnChangeFn<number> = () => {
+  };
+
+  onTouched: OnTouchedFn = () => {
+  };
+
+  readonly UNITS: typeof BarRatingUnitStateEnum = BarRatingUnitStateEnum;
+
   disabled: boolean;
 
   /** Current rating. Can be a decimal value like 3.14 */
-  @Input() rate;
+  rate: ModelSignal<number> = model<number>();
 
   /** Maximal rating that can be given using this widget. */
-  @Input() max = 5;
+  max: InputSignalWithTransform<number, number | string> = input<number, string | number>(this.defaultOptions.maxValue, {
+    transform: numberAttribute
+  });
 
   /** A flag indicating if rating can be updated. */
-  @Input() readOnly = false;
+  readonly: InputSignalWithTransform<boolean, string | boolean> = input<boolean, string | boolean>(this.defaultOptions.readonly, {
+    transform: booleanAttribute,
+    alias: 'readonly'
+  });
 
   /** Set the theme */
-  @Input() theme = 'default';
+  theme: InputSignal<string> = input<string>(this.defaultOptions.theme);
 
   /** Show rating title */
-  @Input() showText = false;
+  showText: InputSignalWithTransform<boolean, string | boolean> = input<boolean, string | boolean>(this.defaultOptions.showText, {
+    transform: booleanAttribute
+  });
 
   /** Replace rate value with a title */
-  @Input() titles = [];
+  titles: InputSignal<string[]> = input<string[]>([]);
 
   /** A flag indicating if rating is required for form validation. */
-  @Input() required = false;
+  required: InputSignalWithTransform<boolean, string | boolean> = input<boolean, string | boolean>(false, { transform: booleanAttribute });
+
+  tabIndex: InputSignalWithTransform<number, number | string> = input<number, string | number>(0, { transform: numberAttribute });
+
+  hoveredIndex: WritableSignal<number> = signal(null);
+
+  contexts: Signal<BarRatingUnitState[]> = computed(() => {
+    const length: number = this.max();
+    const currentRate: number = this.rate();
+    const hovered: number = this.hoveredIndex();
+
+    return Array.from({ length }, (_: unknown, i: number) => {
+      if (hovered) {
+        return i + 1 <= hovered ? BarRatingUnitStateEnum.active : BarRatingUnitStateEnum.inactive;
+      }
+      if (i + 1 <= currentRate) {
+        return BarRatingUnitStateEnum.selected;
+      }
+      if ((i + 1 === Math.round(currentRate) && currentRate % 1) >= 0.5) {
+        return BarRatingUnitStateEnum.fraction;
+      }
+      return BarRatingUnitStateEnum.inactive;
+    });
+  });
+
+  ratingText: Signal<string | number> = computed(() => {
+    const value: string | number = this.hoveredIndex() || this.rate();
+    return this.titles()[value] || value;
+  });
 
   /**
-   * An event fired when a user is hovering over a given rating.
-   * Event's payload equals to the rating being hovered over.
+   * A stream that forwards a bar rating click since clicks are not propagated
    */
-  @Output() hover = new EventEmitter<number>();
+  barClick: OutputEmitterRef<number> = output<number>();
+
+  customActiveRating: Signal<ActiveRating> = contentChild(ActiveRating);
+  customInActiveRating: Signal<InactiveRating> = contentChild(InactiveRating);
+  customFractionRating: Signal<FractionRating> = contentChild(FractionRating);
+
+  updateRating(value: number): void {
+    this.rate.set(value);
+    this.onChange(value);
+  }
 
   /**
-   * An event fired when a user stops hovering over a given rating.
-   * Event's payload equals to the rating of the last item being hovered over.
+   * This is the initial value set to the component
    */
-  @Output() leave = new EventEmitter<number>();
-
-  /**
-   * An event fired when a user selects a new rating.
-   * Event's payload equals to the newly selected rating.
-   */
-  @Output() rateChange = new EventEmitter<number>(true);
-
-  constructor(private changeDetectorRef: ChangeDetectorRef) {
-  }
-
-  ngOnChanges(changes: SimpleChanges): void {
-    if (changes.rate) {
-      this.update(this.rate);
-    }
-  }
-
-  ngOnInit(): void {
-    this.contexts = Array.from({ length: this.max }, (context, i) => ({
-      selected: false,
-      fraction: false,
-      active: false,
-      click: (e) => this.handleClick(e, i + 1),
-      enter: () => this.handleEnter(i + 1)
-    }));
-
-    this.updateState(this.rate);
-  }
-
-  update(newRate: number, internalChange = true): void {
-    if (!this.readOnly && !this.disabled && this.rate !== newRate) {
-      this.rate = newRate;
-      this.rateChange.emit(this.rate);
-    }
-    if (internalChange) {
-      this.onChange(this.rate);
-      this.onTouched();
-    }
-    this.updateState(this.rate);
-  }
-
-  /** Reset rate value */
-  reset(): void {
-    this.leave.emit(this.nextRate);
-    this.updateState(this.rate);
-  }
-
-  private updateState(nextValue): void {
-    /** Set rate value as text */
-    this.nextRate = nextValue - 1;
-    /** Set the rating */
-    this.contexts = Array.from({ length: this.max }, (context, index) => ({
-      selected: index + 1 <= nextValue,
-      fraction: (index + 1 === Math.round(nextValue) && nextValue % 1) >= 0.5,
-      active: false,
-      click: (e) => this.handleClick(e, index),
-      enter: () => this.handleEnter(index)
-    }));
-  }
-
-  private handleClick(e, value): void {
-    /** (NOT TESTED) Remove 300ms click delay on touch devices */
-    e.preventDefault();
-    e.stopPropagation();
-    this.update(value + 1);
-  }
-
-  private handleEnter(index): void {
-    if (!this.disabled && !this.readOnly) {
-      /** Add selected class for rating hover */
-      this.contexts.map((context, i) => {
-        context.active = i <= index;
-        context.fraction = false;
-        context.selected = false;
-      });
-      this.nextRate = index;
-      this.hover.emit(index);
-    }
-  }
-
-  /** This is the initial value set to the component */
   writeValue(value: number): void {
-    this.update(value, false);
-    this.changeDetectorRef.markForCheck();
+    if (value !== null) {
+      this.rate.set(value);
+    }
   }
 
-  validate(c: FormControl): { required: boolean } | null {
-    return (this.required && !c.value) ? { required: true } : null;
+  validate(c: UntypedFormControl): { required: boolean } | null {
+    return (this.required() && !c.value) ? { required: true } : null;
   }
 
-  onChange = (_: any) => {
-  };
-  onTouched = () => {
-  };
-
-  registerOnChange(fn: (value: any) => any): void {
+  registerOnChange(fn: OnChangeFn<number>): void {
     this.onChange = fn;
   }
 
-  registerOnTouched(fn: () => any): void {
+  registerOnTouched(fn: () => OnTouchedFn): void {
     this.onTouched = fn;
   }
 
@@ -177,4 +164,51 @@ export class BarRating implements OnInit, OnChanges, ControlValueAccessor, Valid
     this.disabled = isDisabled;
   }
 
+  handleKeydown(event: KeyboardEvent): void {
+    if (this.readonly() || this.disabled) return;
+
+    const currentRating: number = this.rate();
+    let newRating: number = currentRating;
+
+    switch (event.key) {
+      case 'ArrowRight':
+        newRating = Math.min(currentRating + 1, this.max()); // Increase rating
+        event.preventDefault();
+        break;
+      case 'ArrowLeft':
+        newRating = Math.max(currentRating - 1, 1); // Decrease rating
+        event.preventDefault();
+        break;
+      case 'Enter':
+      case ' ':
+        // Select the current rating (handled in click logic)
+        event.preventDefault();
+        break;
+      default:
+        return; // Ignore other keys
+    }
+
+    if (newRating !== currentRating) {
+      this.updateRating(newRating);
+      this.barClick.emit(newRating);
+      this.hoveredIndex.set(newRating);
+    }
+  }
+
+  handleClick(event: MouseEvent): void {
+    if (this.readonly() || this.disabled) return;
+
+    const target = event.target as HTMLElement;
+    const unitElement = target.closest('.br-unit') as HTMLElement;
+    const value: string = unitElement?.getAttribute('data-value');
+
+    if (value) {
+      const rating: number = parseInt(value, 10);
+      this.updateRating(rating);
+      this.barClick.emit(rating);
+    }
+  }
 }
+
+type OnChangeFn<T> = (value: T) => void;
+type OnTouchedFn = () => void;